Move all browser and page related files into a browser/ directory
